requirements.txt
google-cloud-core==2.4.1
google-api-core==2.21.0
dbt-core==1.8.7
dbt-bigquery==1.8.3
